backend/requirements.txt
# Фреймворк и плагины
Django==5.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9

drf-spectacular==0.27.2
django-filter==24.2
django-cors-headers==4.3.1
drf_standardized_errors==0.14.1
gunicorn==21.2.0
requests==2.31.0
ipython==8.23.0
ruff==0.4.4
pillow==10.2.0
beautifulsoup4==4.12.3
